engine: move MVCCStats tests to own file
Release note: None
